requirements.txt
celery==4.0.2
Django==1.11.11
django-redis-cache==1.7.1
djangorestframework==3.7.7
openpyxl==2.5.1
redis==2.10.5
